Removing skinned assets from project before switching
